requirements.txt
influxdb-client
pyairtouch
python-dotenv